Reorganized packages and added classes that help with implementation.
