+++ app/src/main/java/edu/uw/tcss450/group4/weatherchatapp/io/RequeestQueueSingleton.java
@@ -1,0 +1,4 @@
+package edu.uw.tcss450.group4.weatherchatapp.io;
+
+public class RequeestQueueSingleton {
+}
