+++ app/src/main/java/edu/uw/tcss450/group4/weatherchatapp/ui/signin/SignInViewModel.java
@@ -1,4 +1,4 @@
-package edu.uw.tcss450.group4.weatherchatapp.ui;
+package edu.uw.tcss450.group4.weatherchatapp.ui.signin;
 
 import androidx.lifecycle.ViewModel;
