+++ app/src/main/res/layout/fragment_register.xml
@@ -6,7 +6,7 @@
     android:id="@+id/layout_root"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.RegisterFragment">
+    tools:context=".ui.register.RegisterFragment">
 
     <EditText
         android:id="@+id/edit_email"
